# advertisement_Design_back/src/main/java/com/advertisementdesign/back/controller/ConversationController.java
package com.advertisementdesign.back.controller;

import com.advertisementdesign.back.api.conversation.ConversationModels;
import com.advertisementdesign.back.common.api.Result;
import com.advertisementdesign.back.service.ConversationService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@Tag(name = "Conversation", description = "会话与消息接口")
@RestController
@RequestMapping("/api")
@RequiredArgsConstructor
public class ConversationController {
    private final ConversationService conversationService;

    @Operation(summary = "当前用户会话列表")
    @GetMapping("/conversations")
    public Result<List<ConversationModels.ConversationVO>> list() {
        return Result.success(conversationService.list());
    }

    @Operation(summary = "会话消息列表")
    @GetMapping("/conversations/{conversationId}/messages")
    public Result<ConversationModels.MessageCursorPage> messages(
            @PathVariable Long conversationId,
            @RequestParam(required = false) Long beforeMessageId,
            @RequestParam(defaultValue = "20") long size) {
        return Result.success(conversationService.messages(conversationId, beforeMessageId, size));
    }

    @Operation(summary = "发送消息")
    @PostMapping("/conversations/{conversationId}/messages")
    public Result<ConversationModels.MessageVO> sendMessage(
            @PathVariable Long conversationId,
            @Valid @org.springframework.web.bind.annotation.RequestBody ConversationModels.SendMessageRequest request) {
        return Result.success(conversationService.sendMessage(conversationId, request));
    }

    @Operation(summary = "标记会话已读")
    @PostMapping("/conversations/{conversationId}/read")
    public Result<ConversationModels.ConversationReadStateVO> markRead(
            @PathVariable Long conversationId,
            @Valid @org.springframework.web.bind.annotation.RequestBody ConversationModels.MarkReadRequest request) {
        return Result.success(conversationService.markRead(conversationId, request));
    }

    @Operation(summary = "删除消息")
    @DeleteMapping("/messages/{messageId}")
    public Result<Boolean> deleteMessage(@PathVariable Long messageId) {
        return Result.success(conversationService.deleteMessage(messageId));
    }
}